+++ oex_plugin/apps.py
@@ -5,9 +5,6 @@ oex_plugin Django application initialization.
 from django.apps import AppConfig
 from edx_django_utils.plugins.constants import PluginSettings, PluginURLs
 
-import oex_plugin.override_monkey.rg_xmodule  # NOTE: import is needed to invoke monkey overrides
-
-
 class OexPluginConfig(AppConfig):
     """
     Configuration for the oex_plugin Django application.
@@ -20,6 +17,14 @@ class OexPluginConfig(AppConfig):
     plugin_app = {
         PluginSettings.CONFIG: {
             'lms.djangoapp': {
+                'production': {
+                    PluginSettings.RELATIVE_PATH: 'settings.production',
+                },
+                'devstack': {
+                    PluginSettings.RELATIVE_PATH: 'settings.devstack',
+                },
+            },
+            'cms.djangoapp': {
                 'production': {
                     PluginSettings.RELATIVE_PATH: 'settings.production',
                 },